installed sass, created sass structure, moved App component to components, installed router
