+++ src/components/App/App.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import './App.scss';
+
+const App: React.FC = () => {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <h1>App is working</h1>
+      </header>
+    </div>
+  );
+}
+
+export default App;
