+++ src/index.scss
@@ -1,0 +1,7 @@
+@import './sass/abstracts/mixins';
+@import './sass/abstracts/variables';
+@import './sass/base/reset';
+@import './sass/base/typography';
+@import './sass/componentsSass/buttons';
+@import './sass/layout/navigation';
+@import './sass/layout/footer';
