+++ src/sass/abstracts/_variables.scss
@@ -1,0 +1,1 @@
+$base-color: #c6538c;
